refactor: rename cc-by-sa.js to selection-utils.js and update imports
