+++ app/scripts/selection-utils.js
